Get the api catalogue

Get the api catalogue

api_catalogue()

Details

Get the catalogue of api and store it as a list.

Examples

api_catalogue()

#> [[1]]
#> Api: api.scb.se ('scb')
#>      Statistics Sweden 
#> Version(s)   : v1 
#> Language(s)  : en, sv 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://api.scb.se/OV0104/[version]/doris/[lang] 
#> 
#> [[2]]
#> Api: pxnet2.stat.fi ('statfi')
#>      Statistics Finland 
#> Version(s)   : v1 
#> Language(s)  : fi, sv, en 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://pxnet2.stat.fi/PXWeb/api/[version]/[lang] 
#> 
#> [[3]]
#> Api: pxwebapi2.stat.fi
#>      Statistics Finland (old version) 
#> Version(s)   : v1 
#> Language(s)  : fi 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://pxwebapi2.stat.fi/PXWeb/api/[version]/[lang] 
#> 
#> [[4]]
#> Api: statistik.sjv.se ('jordbruksverket')
#>      The Swedish agricultural authority 
#> Version(s)   : v1 
#> Language(s)  : sv 
#> Limit(s)     : 10 calls per 10 sec.
#>                1000  values per call.
#> Url template :
#>  http://statistik.sjv.se/PXWeb/api/[version]/[lang] 
#> 
#> [[5]]
#> Api: pxweb.orebro.se
#>      Orebro municipality in Sweden 
#> Version(s)   : v1 
#> Language(s)  : sv 
#> Limit(s)     : 10 calls per 10 sec.
#>                1000  values per call.
#> Url template :
#>  http://pxweb.orebro.se/api/[version]/[lang] 
#> 
#> [[6]]
#> Api: fohm-app.folkhalsomyndigheten.se ('fohm')
#>      The Public Health Agency of Sweden 
#> Version(s)   : v1 
#> Language(s)  : en, sv 
#> Limit(s)     : 10 calls per 10 sec.
#>                1000  values per call.
#> Url template :
#>  http://fohm-app.folkhalsomyndigheten.se/Folkhalsodata/api/[version]/[lang] 
#> 
#> [[7]]
#> Api: statistik.konj.se ('konj')
#>      The Swedish national institute of economic research 
#> Version(s)   : v1 
#> Language(s)  : en, sv 
#> Limit(s)     : 30 calls per 1 sec.
#>                10000  values per call.
#> Url template :
#>  http://statistik.konj.se/PXWeb/api/[version]/[lang] 
#> 
#> [[8]]
#> Api: prognos.konj.se ('konjforcast')
#>      The Swedish national institute of economic research, forecast database 
#> Version(s)   : v1 
#> Language(s)  : sv 
#> Limit(s)     : 30 calls per 10 sec.
#>                10000  values per call.
#> Url template :
#>  http://prognos.konj.se/PXWeb/api/[version]/[lang] 
#> 
#> [[9]]
#> Api: statdb.luke.fi ('luke')
#>      LUKE Natural Resources Institute Finland 
#> Version(s)   : v1 
#> Language(s)  : fi, sv, en 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://statdb.luke.fi/PXWeb/api/[version]/[lang] 
#> 
#> [[10]]
#> Api: vero2.stat.fi ('vero')
#>      Verohallinto - Finnish Tax Administration 
#> Version(s)   : v1 
#> Language(s)  : fi, sv, en 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://vero2.stat.fi/PXWeb/api/[version]/[lang] 
#> 
#> [[11]]
#> Api: px.hagstofa.is ('statice')
#>      Statistics Iceland 
#> Version(s)   : v1 
#> Language(s)  : is, en 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://px.hagstofa.is/px[lang]/api/[version]/[lang] 
#> 
#> [[12]]
#> Api: statistik.helsingborg.se ('helsingborg')
#>      Helsingborg town in Sweden 
#> Version(s)   : v1 
#> Language(s)  : sv 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://statistik.helsingborg.se/PXWeb/api/[version]/[lang] 
#> 
#> [[13]]
#> Api: statistik.linkoping.se ('linkoping')
#>      Linkoping municipality in Sweden 
#> Version(s)   : v1 
#> Language(s)  : sv, en 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://statistik.linkoping.se/PXWeb/api/[version]/[lang] 
#> 
#> [[14]]
#> Api: pxwebb2014.vgregion.se ('vgregion')
#>      Vastra Gotaland Region in Sweden 
#> Version(s)   : v1 
#> Language(s)  : sv, en 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://pxwebb2014.vgregion.se/pxweb/api/[version]/[lang] 
#> 
#> [[15]]
#> Api: bank.stat.gl ('greenland')
#>      Statbank Greenland 
#> Version(s)   : v1 
#> Language(s)  : kl, da, en 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://bank.stat.gl/api/[version]/[lang] 
#> 
#> [[16]]
#> Api: px.rsv.is/ ('rsv')
#>      Icelandic Centre for Retail Studies 
#> Version(s)   : v1 
#> Language(s)  : is, en 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://px.rsv.is/PXWeb/api/[version]/[lang] 
#> 
#> [[17]]
#> Api: statbank.hagstova.fo ('hagstovan')
#>      Statistics Faroe Islands 
#> Version(s)   : v1 
#> Language(s)  : en, fo 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://statbank.hagstova.fo/api/[version]/[lang] 
#> 
#> [[18]]
#> Api: data.ssb.no ('ssb')
#>      Statistics Norway 
#> Version(s)   : v0 
#> Language(s)  : en 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://data.ssb.no/api/[version]/[lang] 
#> 
#> [[19]]
#> Api: pxweb.asub.ax ('asub')
#>      Statistics Aland 
#> Version(s)   : v1 
#> Language(s)  : sv, en 
#> Limit(s)     : 30 calls per 10 sec.
#>                1e+05  values per call.
#> Url template :
#>  http://pxweb.asub.ax/PXWeb/api/[version]/[lang]/ 
#> 
#> [[20]]
#> Api: makstat.stat.gov.mk ('makstat')
#>      State Statistical Office of the Republic of Macedonia 
#> Version(s)   : v1 
#> Language(s)  : en, mk 
#> Limit(s)     : 10 calls per 10 sec.
#>                1000  values per call.
#> Url template :
#>  http://makstat.stat.gov.mk/PXWeb/api/[version]/[lang]/MakStat/ 
#> 
#> [[21]]
#> Api: data.csb.gov.lv ('csb_lv')
#>      Statistics Latvia 
#> Version(s)   : v1 
#> Language(s)  : en, lv 
#> Limit(s)     : 10 calls per 10 sec.
#>                1000  values per call.
#> Url template :
#>  http://data.csb.gov.lv/api/[version]/[lang]/ 
#>