Interface to PX-WEB APIs from R

PXWEB is a common web API used by many European Statistical agencies to disseminate official statistics. The pxweb package facilitates connections and usage of these APIs.

References

The decription of the PXWEB API here: https://www.scb.se/en/About-us/about-the-website-and-terms-of-use/open-data-api/ The official home page of PXWEB can be found here: https://www.scb.se/en/services/statistical-programs-for-px-files/px-web/

See also

Author

Maintainer: Mans Magnusson mons.magnusson@gmail.com (ORCID)

Authors:

Markus Kainu

Janne Huovari

Leo Lahti (ORCID)

Other contributors:

Love Hansson [contributor]

Eydun Nielsen [contributor]

Bo Werth [contributor]

Thomas Runarsson [contributor]

Torbjörn Lindquist [contributor]

Palmar Thorsteinsson [contributor]

Pyry Kantanen [contributor]

Sebastian Ankargren [contributor]