Coerce a pxweb_data object to a data.frame

Coerce a pxweb_data object to a data.frame

Usage

pxweb_as_data_frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE,
  column.name.type = "text",
  variable.value.type = "text"
)

# S3 method for class 'pxweb_data'
pxweb_as_data_frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE,
  column.name.type = "text",
  variable.value.type = "text"
)

# S3 method for class 'pxweb_data_v2'
pxweb_as_data_frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE,
  column.name.type = "text",
  variable.value.type = "text"
)

# S3 method for class 'pxweb_data_comments'
pxweb_as_data_frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE
)

# S3 method for class 'pxweb_data_comment'
pxweb_as_data_frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE
)

# S3 method for class 'pxweb_levels'
pxweb_as_data_frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE
)

# S3 method for class 'pxweb_data'
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE,
  column.name.type = "text",
  variable.value.type = "text"
)

# S3 method for class 'pxweb_data_v2'
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE,
  column.name.type = "text",
  variable.value.type = "text"
)

# S3 method for class 'pxweb_data_comments'
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE
)

# S3 method for class 'pxweb_levels'
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE
)

# S3 method for class 'pxweb_metadata'
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE
)

pxweb_as_matrix(
  x,
  row.names = NULL,
  column.name.type = "text",
  variable.value.type = "text"
)

# S3 method for class 'pxweb_data'
pxweb_as_matrix(
  x,
  row.names = NULL,
  column.name.type = "text",
  variable.value.type = "text"
)

# S3 method for class 'pxweb_data_v2'
pxweb_as_matrix(
  x,
  row.names = NULL,
  column.name.type = "text",
  variable.value.type = "text"
)

# S3 method for class 'pxweb_data'
as.matrix(
  x,
  ...,
  row.names = NULL,
  column.name.type = "text",
  variable.value.type = "text"
)

# S3 method for class 'pxweb_data_v2'
as.matrix(
  x,
  ...,
  row.names = NULL,
  column.name.type = "text",
  variable.value.type = "text"
)

pxweb_pxd_slot_idx_pos(x)

Arguments

x

an object to convert to data.frame.

row.names

See as.data.frame.

optional

See as.data.frame.

...

See as.data.frame.

stringsAsFactors

See as.data.frame.

column.name.type

character: should code or text be used as column names?

variable.value.type

character: should code or text be used as values in columns?

See also